+++ src/plugins/dev-mode/check-schema.ts
@@ -182,7 +182,14 @@ function getSchemaPropertyRealPath(shortPath: string) {
  * @throws {Error} if something is not ok
  */
 export function checkSchema(jsonSchema: RxJsonSchema) {
-    // check _rev
+
+    if (!jsonSchema.hasOwnProperty('properties')) {
+        throw newRxError('SC29', {
+            schema: jsonSchema
+        });
+    }
+
+    // _rev MUST NOT exist, it is added by RxDB
     if (jsonSchema.properties._rev) {
         throw newRxError('SC10', {
             schema: jsonSchema